+++ CHANGELOG.md
@@ -9,15 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Specify docker image when publishing a module
 - Improvements for AsyncConnector.
 - Improvements for Async Http workflow
 - Remove duplicated parts and make the code more uniform for async http workflow
+
+## 1.16.1 - 2024-10-30
+
+### Changed
+
+- Specify docker image when publishing a module
+- Move from error to info the message when no event was collected from severals seconds
 
 ### Fixed
 
 - Replace ulrllib.parse.urljoin by posixpath.join in AsyncConnector
 - Fix tests for async version of connector.
+- Fix support for boolean arguments in the json_argument method
 
 ## 1.16.0 - 2024-10-16